+++ Quoin.xcodeproj/project.xcworkspace/contents.xcworkspacedata
