+++ Quoin.xcodeproj/project.xcworkspace/xcshareddata/IDEWorkspaceChecks.plist
