+++ Quoin/Supporting Files/Quoin.h
@@ -1,0 +1,9 @@
+#import <Foundation/Foundation.h>
+
+//! Project version number for Quoin.
+FOUNDATION_EXPORT double QuoinVersionNumber;
+
+//! Project version string for Quoin.
+FOUNDATION_EXPORT const unsigned char QuoinVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <Quoin/PublicHeader.h>
