--- RedUx/Supporting Files/RedUx.h
@@ -1,18 +1,0 @@
-//
-//  RedUx.h
-//  RedUx
-//
-//  Created by Red Davis on 16/06/2021.
-//
-
-#import <Foundation/Foundation.h>
-
-//! Project version number for RedUx.
-FOUNDATION_EXPORT double RedUxVersionNumber;
-
-//! Project version string for RedUx.
-FOUNDATION_EXPORT const unsigned char RedUxVersionString[];
-
-// In this header, you should import all the public headers of your framework using statements like #import <RedUx/PublicHeader.h>
-
-
